Split step manifest state and store helpers
