+++ modal_app/step_manifests/__init__.py
@@ -1,1 +1,1 @@
-"""Step-manifest runtime helpers for Modal pipeline orchestration."""
+"""Step-manifest state and persistence helpers for Modal orchestration."""
