+++ modal_app/step_manifests/store.py
@@ -1,23 +1,17 @@
-"""Runtime helpers for Modal pipeline step manifests."""
+"""Persistence helpers for Modal pipeline step manifests."""
 
 from __future__ import annotations
 
 import json
-import os
 import time
-from dataclasses import asdict, dataclass, field, fields
 from datetime import datetime, timezone
-from pathlib import Path
-from typing import Any, Optional
+from typing import Any
 
-from policyengine_us_data.utils.run_context import RunContext
 from policyengine_us_data.utils.step_manifest import (
     ArtifactReference,
     ReuseMeasurement,
     RunManifest,
     StepManifest,
-    collect_artifacts,
-    collect_directory_artifacts,
     evaluate_step_reuse,
     read_step_manifest,
     run_manifest_path,
@@ -27,108 +21,12 @@ from policyengine_us_data.utils.step_manifest import (
     write_step_manifest,
 )
 
-PIPELINE_MOUNT = "/pipeline"
-STAGING_MOUNT = "/staging"
-ARTIFACTS_BASE = f"{PIPELINE_MOUNT}/artifacts"
-RUNS_DIR = f"{PIPELINE_MOUNT}/runs"
-
-RUN_STEP_IDS = [
-    "01_build_datasets",
-    "02_build_package",
-    "03_fit_weights_regional",
-    "03_fit_weights_national",
-    "04_build_h5_regional",
-    "04_build_h5_national",
-    "04_stage_base_datasets",
-    "04_upload_diagnostics",
-    "05_promote_release",
-]
+from modal_app.step_manifests.state import RUN_STEP_IDS, RunMetadata, run_dir
 
 
-def artifacts_dir_for_run(run_id: str) -> str:
-    """Return the run-scoped artifacts directory."""
-    if run_id:
-        return f"{ARTIFACTS_BASE}/{run_id}"
-    return ARTIFACTS_BASE
-
-
-@dataclass
-class RunMetadata:
-    """Metadata for a pipeline run."""
-
-    run_id: str
-    branch: str
-    sha: str
-    version: str
-    start_time: str
-    status: str
-    step_timings: dict = field(default_factory=dict)
-    error: Optional[str] = None
-    resume_history: list = field(default_factory=list)
-    fingerprint: Optional[str] = None
-    regional_fingerprint: Optional[str] = None
-    run_context: dict = field(default_factory=dict)
-    modal_app_name: Optional[str] = None
-    modal_environment: Optional[str] = None
-    hf_staging_prefix: Optional[str] = None
-
-    def __post_init__(self) -> None:
-        if self.regional_fingerprint is None and self.fingerprint is not None:
-            self.regional_fingerprint = self.fingerprint
-        if self.fingerprint is None and self.regional_fingerprint is not None:
-            self.fingerprint = self.regional_fingerprint
-
-    def to_dict(self) -> dict:
-        data = asdict(self)
-        if (
-            data.get("fingerprint") is None
-            and data.get("regional_fingerprint") is not None
-        ):
-            data["fingerprint"] = data["regional_fingerprint"]
-        return data
-
-    @classmethod
-    def from_dict(cls, data: dict) -> "RunMetadata":
-        data = dict(data)
-        if "run_context" not in data and "publication_context" in data:
-            data["run_context"] = data["publication_context"]
-        if (
-            data.get("regional_fingerprint") is None
-            and data.get("fingerprint") is not None
-        ):
-            data["regional_fingerprint"] = data["fingerprint"]
-        allowed_fields = {field.name for field in fields(cls)}
-        return cls(
-            **{key: value for key, value in data.items() if key in allowed_fields}
-        )
-
-
-def apply_run_context_env(context: RunContext) -> None:
-    """Expose run context to subprocess upload helpers."""
-    for key, value in context.export_env().items():
-        os.environ[key] = value
-
-
-def metadata_run_fields(context: RunContext) -> dict:
-    return {
-        "run_context": context.to_dict(),
-        "modal_app_name": context.modal_app_name,
-        "modal_environment": context.modal_environment,
-        "hf_staging_prefix": context.hf_staging_prefix,
-    }
-
-
-def run_dir(run_id: str) -> Path:
-    return Path(RUNS_DIR) / run_id
-
-
-def artifacts_dir(run_id: str) -> Path:
-    return Path(artifacts_dir_for_run(run_id))
-
-
-def _write_run_manifest(meta: RunMetadata) -> None:
-    """Write the run-scoped execution ledger."""
-    manifest = RunManifest(
+def build_run_manifest(meta: RunMetadata) -> RunManifest:
+    """Build the run-scoped execution ledger from compatibility metadata."""
+    return RunManifest(
         run_id=meta.run_id,
         branch=meta.branch,
         sha=meta.sha,
@@ -147,7 +45,14 @@ def _write_run_manifest(meta: RunMetadata) -> None:
         resume_history=meta.resume_history,
         error=meta.error,
     )
-    write_run_manifest(run_manifest_path(run_dir(meta.run_id)), manifest)
+
+
+def write_run_manifest_for_meta(meta: RunMetadata) -> None:
+    """Write the canonical run manifest for a pipeline run."""
+    write_run_manifest(
+        run_manifest_path(run_dir(meta.run_id)),
+        build_run_manifest(meta),
+    )
 
 
 def write_run_meta(meta: RunMetadata, vol: Any) -> None:
@@ -157,7 +62,7 @@ def write_run_meta(meta: RunMetadata, vol: Any) -> None:
     meta_path = destination / "meta.json"
     with open(meta_path, "w") as f:
         json.dump(meta.to_dict(), f, indent=2)
-    _write_run_manifest(meta)
+    write_run_manifest_for_meta(meta)
     vol.commit()
 
 
@@ -169,12 +74,6 @@ def read_run_meta(run_id: str, vol: Any) -> RunMetadata:
         raise FileNotFoundError(f"No metadata found for run {run_id} at {meta_path}")
     with open(meta_path) as f:
         return RunMetadata.from_dict(json.load(f))
-
-
-def step_completed(meta: RunMetadata, step: str) -> bool:
-    """Check if a legacy step is marked completed in compatibility metadata."""
-    timing = meta.step_timings.get(step, {})
-    return timing.get("status") == "completed"
 
 
 def _next_step_attempt(run_id: str, step_id: str) -> int:
@@ -328,51 +227,6 @@ def step_reusable(
         expected_input_identities=expected_input_identities,
         expected_parameters=expected_parameters,
     )
-
-
-def artifact_identity(path: str | Path) -> dict:
-    artifact = ArtifactReference.from_path(path)
-    return {
-        "path": artifact.path,
-        "size_bytes": artifact.size_bytes,
-        "sha256": artifact.sha256,
-    }
-
-
-def artifact_identities(paths: dict[str, str | Path]) -> dict:
-    identities = {}
-    for label, path in paths.items():
-        artifact_path = Path(path)
-        identities[label] = (
-            artifact_identity(artifact_path)
-            if artifact_path.exists()
-            else {"path": str(artifact_path), "missing": True}
-        )
-    return identities
-
-
-def collect_diagnostics(run_id: str) -> list[ArtifactReference]:
-    return collect_directory_artifacts(
-        run_dir(run_id) / "diagnostics",
-        patterns=("*.csv", "*.json", "*.txt"),
-        role="diagnostic",
-    )
-
-
-def collect_staging_outputs(run_id: str, *, scope: str) -> list[ArtifactReference]:
-    scoped_run_dir = Path(STAGING_MOUNT) / run_id
-    paths: list[Path] = []
-    if scope == "regional":
-        for subdir in ("states", "districts", "cities"):
-            paths.extend(sorted((scoped_run_dir / subdir).glob("*.h5")))
-        manifest_path = scoped_run_dir / "manifest.json"
-        if manifest_path.exists():
-            paths.append(manifest_path)
-    elif scope == "national":
-        paths.extend(sorted((scoped_run_dir / "national").glob("*.h5")))
-    else:
-        raise ValueError(f"Unknown H5 output scope: {scope}")
-    return collect_artifacts(paths, missing_ok=True)
 
 
 def record_step(
